feat: add sounds to correct/incorrect answers, add color variables and GameLogicController
