+++ src/app/sprint-game/SprintService.ts
@@ -1,11 +1,12 @@
 import { Word } from './abstracts';
 
 export default class SprintService {
-  async getWords(level: string, page: string): Promise<Word[]> {
+  static wordCollection: Word[] = [];
+
+  async getWords(level: string, page: string): Promise<void> {
     try {
       const response = await fetch(`https://rslang-leanwords.herokuapp.com/words?page=${page}&group=${level}`);
-      const data = await response.json();
-      return data;
+      SprintService.wordCollection = await response.json();
     } catch (error) {
       throw new Error(`Something went wrong... ${error}`);
     }
